allow partial scan at beginning
