+++ server/tests/test_handler.py
@@ -10,3 +10,4 @@ def test_basic():
     t0 = datetime(2023, 1, 1, 0, 0, 0, tzinfo=timezone.utc)
     t1 = datetime(2023, 1, 2, 0, 0, 10, tzinfo=timezone.utc)
     cmds = handler.basic_policy(t0, t1, {})
+    assert len(cmds) == 4194
