+++ src/schedlib/policy.py
@@ -1,5 +1,6 @@
-from dataclasses import dataclass
 from typing import FrozenSet
+from chex import dataclass
+
 
 @dataclass(frozen=True)
 class Rule: 
@@ -11,4 +12,16 @@ class RuleSet:
 
 @dataclass(frozen=True)
 class Policy:
-    pass
+    config: dict
+
+    @classmethod
+    def from_config(cls, config: dict):
+        return cls(config=config)
+
+    def get(self, *args, **kwargs):
+        return self.config.get(*args, **kwargs)
+
+@dataclass(frozen=True)
+class BasicPolicy(Policy):
+    def get_block_tolarance(self, block):
+        return self.get('block_tolarance', {}).get(str(type(block)), 0)
